Update functional tests
