+++ x-pack/test/functional/apps/spaces/create_edit_space/create_edit_space.ts
@@ -15,6 +15,7 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
   const testSubjects = getService('testSubjects');
   const spacesServices = getService('spaces');
   const log = getService('log');
+  const find = getService('find');
 
   describe('Spaces Management: Create and Edit', () => {
     before(async () => {
@@ -43,6 +44,7 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
 
       it('create a space with a given name', async () => {
         await PageObjects.spaceSelector.addSpaceName(spaceName);
+        await PageObjects.spaceSelector.changeSolutionView('classic');
         await PageObjects.spaceSelector.clickSaveSpaceCreation();
         await testSubjects.existOrFail(`spacesListTableRow-${spaceId}`);
       });
@@ -60,6 +62,7 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
           name: spaceName,
           disabledFeatures: [],
           color: '#AABBCC',
+          solution: 'classic',
         });
 
         await PageObjects.common.navigateToApp('spacesManagement');
@@ -103,14 +106,100 @@ export default function ({ getPageObjects, getService }: FtrProviderContext) {
     });
 
     describe('solution view', () => {
-      it('does not show solution view panel', async () => {
+      it('does show the solution view panel', async () => {
         await PageObjects.common.navigateToUrl('management', 'kibana/spaces/edit/default', {
           shouldUseHashForSubUrl: false,
         });
 
         await testSubjects.existOrFail('spaces-view-page');
         await testSubjects.existOrFail('spaces-view-page > generalPanel');
-        await testSubjects.missingOrFail('spaces-view-page > navigationPanel'); // xpack.spaces.allowSolutionVisibility is not enabled, so the solution view picker should not appear
+        await testSubjects.existOrFail('spaces-view-page > navigationPanel');
+      });
+
+      it('changes the space solution and updates the side navigation', async () => {
+        await PageObjects.common.navigateToUrl('management', 'kibana/spaces/edit/default', {
+          shouldUseHashForSubUrl: false,
+        });
+
+        // Make sure we are on the classic side nav
+        await testSubjects.existOrFail('mgtSideBarNav');
+        await testSubjects.missingOrFail('searchSideNav');
+
+        // change to Enterprise Search
+        await PageObjects.spaceSelector.changeSolutionView('es');
+        await PageObjects.spaceSelector.clickSaveSpaceCreation();
+        await PageObjects.spaceSelector.confirmModal();
+
+        await find.waitForDeletedByCssSelector('.kibanaWelcomeLogo');
+
+        // Search side nav is loaded
+        await testSubjects.existOrFail('searchSideNav');
+        await testSubjects.missingOrFail('mgtSideBarNav');
+
+        // change back to classic
+        await PageObjects.common.navigateToUrl('management', 'kibana/spaces/edit/default', {
+          shouldUseHashForSubUrl: false,
+        });
+
+        await testSubjects.missingOrFail('space-edit-page-user-impact-warning');
+        await PageObjects.spaceSelector.changeSolutionView('classic');
+        await testSubjects.existOrFail('space-edit-page-user-impact-warning'); // Warn that the change will impact other users
+
+        await PageObjects.spaceSelector.clickSaveSpaceCreation();
+        await PageObjects.spaceSelector.confirmModal();
+
+        await testSubjects.existOrFail('mgtSideBarNav');
+        await testSubjects.missingOrFail('searchSideNav');
+      });
+    });
+
+    describe('API-created Space', () => {
+      before(async () => {
+        await spacesServices.create({
+          id: 'foo-space',
+          name: 'Foo Space',
+          disabledFeatures: [],
+          color: '#AABBCC',
+        });
+      });
+
+      after(async () => {
+        await spacesServices.delete('foo-space');
+      });
+
+      it('enabled features can be changed while the solution view remains unselected', async () => {
+        const securityFeatureCheckboxId = 'featureCategoryCheckbox_securitySolution';
+
+        await PageObjects.common.navigateToUrl('management', 'kibana/spaces/edit/foo-space', {
+          shouldUseHashForSubUrl: false,
+        });
+
+        await testSubjects.existOrFail('spaces-view-page');
+
+        // ensure security feature is selected by default
+        expect(await testSubjects.isChecked(securityFeatureCheckboxId)).to.be(true);
+
+        // Do not set a solution view first!
+
+        await PageObjects.spaceSelector.toggleFeatureCategoryCheckbox('securitySolution');
+        //
+        // ensure security feature now unselected
+        expect(await testSubjects.isChecked(securityFeatureCheckboxId)).to.be(false);
+
+        await testSubjects.existOrFail('space-edit-page-user-impact-warning');
+
+        await PageObjects.spaceSelector.clickSaveSpaceCreation();
+
+        await testSubjects.click('confirmModalConfirmButton');
+
+        await testSubjects.existOrFail('spaces-view-page');
+
+        await testSubjects.click('foo-space-hyperlink');
+
+        await testSubjects.existOrFail('spaces-view-page');
+
+        // ensure security feature is still unselected
+        expect(await testSubjects.isChecked(securityFeatureCheckboxId)).to.be(false);
       });
     });
   });
