+++ x-pack/test/functional/apps/spaces/index.ts
@@ -15,5 +15,6 @@ export default function spacesApp({ loadTestFile }: FtrProviderContext) {
     loadTestFile(require.resolve('./enter_space'));
     loadTestFile(require.resolve('./create_edit_space'));
     loadTestFile(require.resolve('./spaces_grid'));
+    loadTestFile(require.resolve('./solution_tour'));
   });
 }
