+++ x-pack/test/functional_solution_sidenav/config.ts
@@ -20,12 +20,7 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
     testFiles: [require.resolve('.')],
     kbnTestServer: {
       ...functionalConfig.get('kbnTestServer'),
-      serverArgs: [
-        ...functionalConfig.get('kbnTestServer.serverArgs'),
-        // Note: the base64 string in the cloud.id config contains the ES endpoint required in the functional tests
-        '--xpack.cloud.id=ftr_fake_cloud_id:aGVsbG8uY29tOjQ0MyRFUzEyM2FiYyRrYm4xMjNhYmM=',
-        '--xpack.cloud.base_url=https://cloud.elastic.co',
-      ],
+      serverArgs: [...functionalConfig.get('kbnTestServer.serverArgs')],
     },
   };
 }
